refactor(capabilities): :recycle: migrate buttons molecules to component-store pattern
- pagination-navigation: pure presentational, no store; fixes relative import from pills-links
- pills-buttons (RedPillButton/BluePillButton): pure presentational, onClick is prop ref not inline lambda
- segmented-control: EffectsStore<{triggerChange}> via useSegmentedControlStore; curry onChange(value)=>()=>void fixes jsx-no-bind
- social-contact: pure presentational
- tag: pure presentational

Each component nested in own kebab folder under buttons/ category dir with index.ts barrel.
