+++ src/components/design-system/molecules/buttons/pagination-navigation/index.ts
@@ -1,0 +1,2 @@
+export { PaginationNavigation } from "./pagination-navigation";
+export type { PageNavigationProps } from "./pagination-navigation";
