+++ src/components/design-system/molecules/buttons/pills-buttons/index.ts
@@ -1,0 +1,1 @@
+export { RedPillButton, BluePillButton } from "./pills-buttons";
