+++ src/components/design-system/molecules/buttons/pills-buttons/pills-buttons.tsx
@@ -1,0 +1,21 @@
+"use client";
+
+import { FC, PropsWithChildren } from "react";
+import { BluePill, RedPill } from "@/components/design-system/atoms/effects/pills";
+
+type PillProps = PropsWithChildren<{
+    onClick: () => void;
+    disabled?: boolean;
+}>;
+
+export const RedPillButton: FC<PillProps> = ({ children, onClick, disabled }) => (
+    <button className="bg-transparent border-none cursor-pointer p-0" onClick={onClick} disabled={disabled}>
+        <RedPill>{children}</RedPill>
+    </button>
+);
+
+export const BluePillButton: FC<PillProps> = ({ children, onClick, disabled }) => (
+    <button className="bg-transparent border-none cursor-pointer p-0" onClick={onClick} disabled={disabled}>
+        <BluePill>{children}</BluePill>
+    </button>
+);
