+++ src/components/design-system/molecules/buttons/segmented-control/index.ts
@@ -1,0 +1,2 @@
+export { SegmentedControl } from "./segmented-control";
+export type { SegmentOption } from "./segmented-control";
