+++ src/components/design-system/molecules/buttons/social-contact/index.ts
@@ -1,0 +1,2 @@
+export { SocialContact } from "./social-contact";
+export type { SocialContactProps } from "./social-contact";
