+++ src/components/design-system/molecules/buttons/tag/index.ts
@@ -1,0 +1,2 @@
+export { Tag } from "./tag";
+export type { TagProps } from "./tag";
